Added file exists check

+++ KSTests/ColorTests.vb
@@ -31,7 +31,7 @@ Imports KS
             Dim ExpectedTemplates As New List(Of String)
             Dim ActualTemplates As New List(Of String)
             Dim PathToTestConfig As String = Path.GetFullPath("TestConfig.ini")
-            File.Copy(PathToTestConfig, paths("Configuration"))
+            If Not File.Exists(paths("Configuration")) Then File.Copy(PathToTestConfig, paths("Configuration"))
             ExpectedTemplates.AddRange(colorTemplates)
             For Each Template As String In colorTemplates
                 TemplateSet(Template)
